requirements.txt
instagrapi
anthropic
python-dotenv
rich
streamlit
requests
